# apps/web/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medicly API",
    "version": "1.0.0",
    "description": "Medicly MVP HTTP API. Cookie session: medicly_sid."
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "medicly_sid"
      }
    },
...
  },
  "paths": {
...
    "/v1/doctors/search": {
      "get": {
        "operationId": "getDoctorsSearch",
        "tags": [
          "doctors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cityId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "clinicId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "specialty",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "offline"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "online"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "both"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "date",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "minRating",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "priceMin",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "priceMax",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "rating"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "nearest_slot"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "nextCursor",
                    "items"
                  ],
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "firstName",
                          "lastName",
                          "specialty",
                          "clinicName",
                          "cityName",
                          "clinicId",
                          "cityId",
                          "photoUrl",
                          "supportedFormats",
                          "nearestFreeAt",
                          "basePrice",
                          "promoPrice",
                          "ratingAverage",
                          "reviewCount",
                          "isFavourite"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "specialty": {
                            "type": "string"
                          },
                          "clinicName": {
                            "type": "string"
                          },
                          "cityName": {
                            "type": "string"
                          },
                          "clinicId": {
                            "type": "string"
                          },
                          "cityId": {
                            "type": "string"
                          },
                          "photoUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "supportedFormats": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "offline"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "online"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "both"
                                ]
                              }
                            ]
                          },
                          "nearestFreeAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "basePrice": {
                            "type": "number"
                          },
                          "promoPrice": {
                            "type": "number",
                            "nullable": true
                          },
                          "ratingAverage": {
                            "type": "number"
                          },
                          "reviewCount": {
                            "type": "number"
                          },
                          "isFavourite": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "prefill": {
                      "type": "object",
                      "required": [
                        "cityId",
                        "clinicId"
                      ],
                      "properties": {
                        "cityId": {
                          "type": "string",
                          "nullable": true
                        },
                        "clinicId": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/doctors/{doctorId}": {
      "get": {
        "operationId": "getDoctorById",
        "tags": [
          "doctors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "doctorId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "specialty",
                    "clinicId",
                    "cityId",
                    "clinicName",
                    "cityName",
                    "address",
                    "photoUrl",
                    "yearsPractice",
                    "visitDurationMinutes",
                    "languages",
                    "descriptionUk",
                    "descriptionEn",
                    "bio",
                    "supportedFormats",
                    "basePrice",
                    "promoPrice",
                    "ratingAverage",
                    "reviewCount",
                    "consultationCount",
                    "isFavourite",
                    "reviews"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "specialty": {
                      "type": "string"
                    },
                    "clinicId": {
                      "type": "string"
                    },
                    "cityId": {
                      "type": "string"
                    },
                    "clinicName": {
                      "type": "string"
                    },
                    "cityName": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "photoUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "yearsPractice": {
                      "type": "number"
                    },
                    "visitDurationMinutes": {
                      "type": "number"
                    },
                    "languages": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "uk"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "en"
                            ]
                          }
                        ]
                      }
                    },
                    "descriptionUk": {
                      "type": "string"
                    },
                    "descriptionEn": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string"
                    },
                    "supportedFormats": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "offline"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "online"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "both"
                          ]
                        }
                      ]
                    },
                    "basePrice": {
                      "type": "number"
                    },
                    "promoPrice": {
                      "type": "number",
                      "nullable": true
                    },
                    "ratingAverage": {
                      "type": "number"
                    },
                    "reviewCount": {
                      "type": "number"
                    },
                    "consultationCount": {
                      "type": "number"
                    },
                    "isFavourite": {
                      "type": "boolean"
                    },
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "rating",
                          "text",
                          "patientDisplayName",
                          "createdAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "number"
                          },
                          "text": {
                            "type": "string"
                          },
                          "patientDisplayName": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/doctors/{doctorId}/calendar": {
      "get": {
        "operationId": "getDoctorCalendar",
        "tags": [
          "doctors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "date",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "contextAppointmentId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "doctorId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "doctorId",
                    "zoneAStart",
                    "zoneAEnd",
                    "visitDurationMinutes",
                    "supportedFormats",
                    "days",
                    "slots"
                  ],
                  "properties": {
                    "doctorId": {
                      "type": "string"
                    },
                    "zoneAStart": {
                      "type": "string"
                    },
                    "zoneAEnd": {
                      "type": "string"
                    },
                    "visitDurationMinutes": {
                      "type": "number"
                    },
                    "supportedFormats": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "offline"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "online"
                            ]
                          }
                        ]
                      }
                    },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "date",
                          "flag"
                        ],
                        "properties": {
                          "date": {
                            "type": "string"
                          },
                          "flag": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "has_free"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "full"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "day_off"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "outside_window"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "past"
                                ]
                              }
                            ]
                          }
                        }
                      }
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "startAt",
                          "status"
                        ],
                        "properties": {
                          "startAt": {
                            "type": "string"
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "free"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "taken"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "reserved"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "past"
                                ]
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api",
      "description": "Vite proxy / same-origin /api"
    }
  ],
...
}